Changed the way objects are handled

+++ MLPlusPlus/src/DataTypes/Array.cpp
@@ -1,5 +1,5 @@
 #include "Array.h"
 
-Array::Array(const std::vector<float>& data)
+Array::Array(const std::vector<double>& data)
 	:data(data) {}
